+++ src/cli.ts
@@ -78,7 +78,10 @@ if (args.analyze) (async () => {
 		if (args.tree) {
 			const treeParts: string[] = args.tree.split('.');
 			let nestedData: Record<string, any> = data;
-			for (const part of treeParts) nestedData = nestedData[part];
+			for (const part of treeParts) {
+				if (!nestedData[part]) throw Error(`TraversalError: Key '${part}' cannot be found on output object.`);
+				nestedData = nestedData[part];
+			}
 			console.log(nestedData);
 		}
 		else {