+++ changelog.md
@@ -3,6 +3,7 @@
 ## Next
 - Changed the error thrown when traversing an invalid tree via CLI to contain a more useful description.
 - Fixed heuristics not applying start- and end-of-line markers properly ([#5](https://github.com/Nixinova/Linguist/issues/5)).
+- Fixed heuristic language fallback not being applied properly.
 
 ## 2.0.2
 *2021-09-09*